Backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.0
cohere==4.37
Pillow==10.2.0 